Move and clean few tests
