+++ test/test_add_mode.py
