+++ test/test_cache.py
