+++ test/test_cmdline.py
