--- test/ufc/test_cffi_3d.py
@@ -1,152 +1,0 @@
-# Copyright (C) 2018 Chris Richardson
-#
-# This file is part of FFC (https://www.fenicsproject.org)
-#
-# SPDX-License-Identifier:    LGPL-3.0-or-later
-
-import cffi
-import numpy as np
-import pytest
-
-import ffc.codegeneration.jit
-import ufl
-
-
-def float_to_type(name):
-    """Map a string name to C and NumPy types"""
-    if name == "double":
-        return "double", np.float64
-    elif name == "double complex":
-        return "double _Complex", np.complex128
-    else:
-        raise RuntimeError("Unknown C type for: {}".format(name))
-
-
-@pytest.fixture(scope="module")
-def lagrange_element():
-    """Compile list of Lagrange elements"""
-    cell = ufl.tetrahedron
-    elements = [ufl.FiniteElement("Lagrange", cell, p) for p in range(1, 5)]
-    compiled_elements, module = ffc.codegeneration.jit.compile_elements(elements)
-    return elements, compiled_elements, module
-
-
-@pytest.fixture(scope="module")
-def hexahedral_element():
-    """Compile list of Lagrange elements"""
-    cell = ufl.hexahedron
-    elements = [ufl.FiniteElement("Lagrange", cell, p) for p in range(1, 5)]
-    compiled_elements, module = ffc.codegeneration.jit.compile_elements(elements)
-    return elements, compiled_elements, module
-
-
-def test_dim_degree(lagrange_element):
-    ufl_elements, compiled_elements, module = lagrange_element
-    for e, compiled_e in zip(ufl_elements, compiled_elements):
-        assert compiled_e[0].geometric_dimension == 3
-        assert compiled_e[0].topological_dimension == 3
-        assert e.degree() == compiled_e[0].degree
-
-
-def test_tabulate_reference_dof_coordinates(lagrange_element):
-    ufl_elements, compiled_elements, module = lagrange_element
-    for e, compiled_e in zip(ufl_elements, compiled_elements):
-        tdim = compiled_e[0].topological_dimension
-        space_dim = compiled_e[0].space_dimension
-        X = np.zeros([space_dim, tdim])
-        X_ptr = module.ffi.cast("double *", module.ffi.from_buffer(X))
-        compiled_e[0].tabulate_reference_dof_coordinates(X_ptr)
-
-
-def test_evaluate_reference_basis_hex(hexahedral_element):
-    ufl_elements, compiled_elements, module = hexahedral_element
-    for e, compiled_e in zip(ufl_elements, compiled_elements):
-        space_dim = compiled_e[0].space_dimension
-        X = np.array([[0.0, 0.0, 0.0], [0.5, 0.5, 0.5], [1.0, 0.0, 0.0],
-                      [0.0, 1.0, 0.0], [1.0, 1.0, 1.0]])
-        npoint = X.shape[0]
-        X_ptr = module.ffi.cast("const double *", module.ffi.from_buffer(X))
-        vals = np.zeros([npoint, space_dim])
-        vals_ptr = module.ffi.cast("double *", module.ffi.from_buffer(vals))
-        compiled_e[0].evaluate_reference_basis(vals_ptr, npoint, X_ptr)
-        assert np.isclose(np.sum(vals), npoint)
-
-
-@pytest.mark.parametrize("degree,coords", [(1, np.array([[0, 0, 0], [0, 0, 3],
-                                                         [0, 2, 0], [0, 2, 3],
-                                                         [1, 0, 0], [1, 0, 3],
-                                                         [1, 2, 0], [1, 2, 3]], dtype=np.float64)),
-                                           (2, np.array([[0, 0, 0], [0, 0, 3], [0, 0, 1.5],
-                                                         [0, 2, 0], [0, 2, 3], [0, 2, 1.5],
-                                                         [0, 1, 0], [0, 1, 3], [0, 1, 1.5],
-                                                         [1, 0, 0], [1, 0, 3], [1, 0, 1.5],
-                                                         [1, 2, 0], [1, 2, 3], [1, 2, 1.5],
-                                                         [1, 1, 0], [1, 1, 3], [1, 1, 1.5],
-                                                         [0.5, 0, 0], [0.5, 0, 3], [0.5, 0, 1.5],
-                                                         [0.5, 2, 0], [0.5, 2, 3], [0.5, 2, 1.5],
-                                                         [0.5, 1, 0], [0.5, 1, 3], [0.5, 1, 1.5]], dtype=np.float64))])
-def test_cmap_hex(degree, coords):
-    # Test for first and second order quadrilateral meshes,
-    # assuming FIAT Tensor Product layout of cell.
-
-    cell = ufl.hexahedron
-    e = ufl.VectorElement("Lagrange", cell, degree)
-    mesh = ufl.Mesh(e)
-    compiled_cmap, module = ffc.codegeneration.jit.compile_coordinate_maps([mesh])
-
-    coords_ptr = module.ffi.cast("double *", module.ffi.from_buffer(coords))
-
-    x = np.array([[1 / 3, 3 / 2, 1]], dtype=np.float64)
-    x_ptr = module.ffi.cast("double *", module.ffi.from_buffer(x))
-    X = np.zeros_like(x)
-    X_ptr = module.ffi.cast("double *", module.ffi.from_buffer(X))
-    compiled_cmap[0].compute_physical_coordinates(X_ptr, X.shape[0], x_ptr, coords_ptr)
-
-    assert(np.isclose(X[0, 0], x[0, 0]))
-    assert(np.isclose(X[0, 1], 2 * x[0, 1]))
-    assert(np.isclose(X[0, 2], 3 * x[0, 2]))
-
-
-@pytest.mark.parametrize("mode,expected_result", [
-    ("double", np.array([[0.5, -1 / 6, -1 / 6, -1 / 6],
-                         [-1 / 6, 1 / 6, 0.0, 0.0],
-                         [-1 / 6, 0.0, 1 / 6, 0.0],
-                         [-1 / 6, 0.0, 0.0, 1 / 6]], dtype=np.float64)),
-    ("double complex",
-     np.array(
-         [[0.5 + 0j, -1 / 6 + 0j, -1 / 6 + 0j, -1 / 6 + 0j],
-          [-1 / 6 + 0j, 1 / 6 + 0j, 0.0 + 0j, 0.0 + 0j],
-          [-1 / 6 + 0j, 0.0 + 0j, 1 / 6 + 0j, 0.0 + 0j],
-          [-1 / 6 + 0j, 0.0 + 0j, 0.0 + 0j, 1 / 6 + 0j]],
-         dtype=np.complex128)),
-])
-def test_laplace_bilinear_form_3d(mode, expected_result):
-    cell = ufl.tetrahedron
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
-    a = ufl.inner(ufl.grad(u), ufl.grad(v)) * ufl.dx
-    forms = [a]
-    compiled_forms, module = ffc.codegeneration.jit.compile_forms(forms, parameters={'scalar_type': mode})
-
-    for f, compiled_f in zip(forms, compiled_forms):
-        assert compiled_f.rank == len(f.arguments())
-
-    form0 = compiled_forms[0][0].create_cell_integral(-1)
-
-    c_type, np_type = float_to_type(mode)
-    A = np.zeros((4, 4), dtype=np_type)
-    w = np.array([], dtype=np_type)
-    c = np.array([], dtype=np_type)
-
-    ffi = cffi.FFI()
-    coords = np.array([0.0, 0.0, 0.0,
-                       1.0, 0.0, 0.0,
-                       0.0, 1.0, 0.0,
-                       0.0, 0.0, 1.0], dtype=np.float64)
-    form0.tabulate_tensor(
-        ffi.cast('{type} *'.format(type=c_type), A.ctypes.data),
-        ffi.cast('{type} *'.format(type=c_type), w.ctypes.data),
-        ffi.cast('{type} *'.format(type=c_type), c.ctypes.data),
-        ffi.cast('double *', coords.ctypes.data), ffi.NULL, ffi.NULL)
-
-    assert np.allclose(A, expected_result)
